# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EECU",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/communities": {
            "get": {
                "tags": [
                    "Громади"
                ],
                "summary": "Отримати список громад",
                "description": "Користувач повинен бути авторизований<br><br> Щоб використати токен Headers запиту має містити:\r\n\t *    <br><br> <em>Accept=application/json<br>Authorization=Bearer token_here</em>",
                "operationId": "CommunitiesGet",
                "responses": {
                    "200": {
                        "description": "Буде повернено масив з громадами та хедер таблиць",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Учасники проекту"
                                        },
                                        "headers": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Назва"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "name"
                                                    },
                                                    "sortable": {
                                                        "type": "boolean",
                                                        "example": "true"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "row": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "hromada"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Громада"
                                                    },
                                                    "chief": {
                                                        "type": "string",
                                                        "example": "Керівник громади"
                                                    },
                                                    "contact_person": {
                                                        "type": "string",
                                                        "example": "Контактна персона"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+38 (096) 000 00 00"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "email@gmail.com"
                                                    },
                                                    "lat": {
                                                        "type": "string",
                                                        "example": "42.2323"
                                                    },
                                                    "lng": {
                                                        "type": "string",
                                                        "example": "22.2313"
                                                    },
                                                    "picture": {
                                                        "type": "string",
                                                        "example": "22"
                                                    },
                                                    "eea_member": {
                                                        "type": "integer",
                                                        "example": "0"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/content": {
            "get": {
                "tags": [
                    "Контент"
                ],
                "summary": "Content",
                "description": "Get information about user",
                "operationId": "content",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Name"
                                        },
                                        "role": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nav": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get information<br><br> Headers have to contain:\r\n\t *    <br><br> <em>Accept=application/json<br>Authorization=Bearer token_here</em>",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}